Create bracket-repeats.ly
This is a workaround for creating repeats which are formatted as "1x" and centered within a bracket spanning a measure.  It is based on a modification of the stencil function for the MeasureCounter grob.  This solution only works with measures that are not split across systems and with single-measure repeats.  (Such enhancements are possible, but may necessitate moving the file to the scheme-engravers category.)