predefined-instruments: remove \addlyrics to have proper voice contexts
With \addlyrics, the notes weren't placed in a SopranoVoice,
just in an ordinary Voice (despite using \defaultchild).

This was because \addlyrics says "hey, I need a Voice" right
after SopranoStaff has been created, and gets the demanded Voice.
And then a Bottom context is there, so SopranoVoice isn't created.

Without \addlyrics, we get the expected result.