Add dedicated draft-mode directory
This project will surely have a number of files, so I think it is a good idea to work on it within a directory.

I'm not sure if the 'snippets' repo is the right place for that, but as long as we don't have a better solution I suggest to go that way.