display-slur-control-points: Several Enhancement by Harm
(1)
 the possibility to configure different cross-color and line-color,
using 'stencil-with-color' from /scm/stencil.scm, with the advantage
that it's now possible to insert all kind of color-arguments:
rgb-color, x11-color, predefined ones like red or green etc
(2)
access all curves 'stencil. (Same like David)
(3)
Code to display first and last control-point and the line between
second and third.
This code is commented for now. Though, I thought better to have it in.
(4)
setting the stencil-extents for lines and crosses to empty-interval.
This solves the tie-issue for 2.16.2, but not for 2.17.x (see below)
(5)
Both examples are extended.
The first to test RepeatTie and LaisserVibrerTie
(6)
The second shows the workaround for the tie-issue with 2.17.26.
Also, some 'layer-settings