Update generalized-offsets.ly
Add a warning if user attempts to offset a property having a default value of '(+inf.0 . -inf.0).  Adding offsets to such a pair results in the same value, so no change would be observed.  This could potentially be confusing to the user (who shouldn't need to know the default values in the first place or consider the results of arithmetical operations on infinity).