notation snippets: reorganize
I've removed all duplicated descriptions.  In my opinion it doesn't
make sense to copy-and-paste so much.  I agree that the description
should be visible to the reader, but i believe we should do it in
another way: include both the description and the image from the
example file in something like the README.md.  Unfortunately, this
may not be possible with GitHub's interface - that's something that
we should do when creating a dedicated web frontend.

I also reverted the scaling-stencils snippet to its previous state,
because this particular snippet isn't worth being divided into
definition and example.  And most importantly, the function that
was defined was not generic - it had ly:flag::print hardcoded,
so it wouldn't work with anything else.