annotate: temporarily change version to 2.19
This is to force compilation with current version.
The snippet doesn't run with 2.16.2 or 2.17.3
2.17.27 however does work.

The offending code is in props-alist.
If I replace its body with (list 1 2 3)
(and of course comment any use of the resulting list)
everything works fine, so it isn't because of the
calling convention with \with
(although the error message suggests this).