oll-doc: Read categories and tags from snippets (TMP)
This is a temporary approach to be able to get the used
tags and categories. It's only intended for the process of
converting all the snippets to the new structure, giving
hints to the snippet authors as to what tags to use.

The parsing of tags and categories is extremely hacky,
and the output is only "for now"