oll-doc: Parse header fields of snippets (WIP)
use a dictionary SnippetDefinition.headerFields
Populate this when parsing the headercontent.

For now it only understands fields with one quoted string in a line.
Others will return None.

Another step to be done: initialize mandatory fields with the dict,
so one can walk through that dict and determine missing fields
in the file.