oll-doc: Add MetadataWidget()
A widget (may later be moved to a dialog or dockable panel)
to display (and later edit) snippet metadata.

This commit simply sets up the widget and places it in the
main window. On loading of a snippet it displays the content
of the snippet's metadata in line or text edits.