oll-doc: first version of HTML document (definition)
Make a first working HTML version of a detail page in the
WebView.
Make the generation as generic as possible (can certainly be
improved, though).
Add initial CSS styles, these have to be improved.
Remove the initial textedits for the snippet definition.
Hide the metadata widget (will be reused as an editor).