move index higher in README.md
For people who have just discovered this repository, learning how to
use it or contribute to it is irrelevant until they first establish
that they are somehow interested in the contents.  So move the index
to appear before the "Using this repository" or "Contributing"
sections.

Additionally, rename the index from "Snippet categories" to
"Contents", since not all items are snippets.