use \removeWithTag instead of \keepWithTag.
That method will at least preserve music that the user has tagged
for some other purpose, and it co-exists with \removeWithTag for other
purposes.

Just in case anyone uses the \divisibleStaff with his own tags, and might
not think to include 'together in \keepWithTag score \divisibleStaff ... ,
I added a function to clean up the divI/divII/together tags when we are
done with them.