Installation script for Abraham's fonts (WIP)
This is an initial version of a script that simplifies
the installation of alternative fonts with the help of
symlinks.

Probably it doesn't work on Windows, and Mac support hasn't
been implemented yet. Also it wouldn't work on installations
through a Linux package manager.

The missing implementations should be easily done, but
not by me ...