Make annotation-colors an option
This is somewhat more complex because the colors are
stored in an alist themselves that shows up as an additional
tree level. For retrieval of a color the option path has to be
created dynamically.

TODO: Consider if that may need a more general approach, sth like
a \getSubOption command.