gridly: Remove redundant includes
It is not necessary to include "grid.ily" (newly __main__.ily")
or grid-templates.ily redundantly from the already included
files. Actually that's even a problem because the file would be
parsed each time.
(This is different from e.g. Python where you have to import
a module to make its interface available.)