scholarly: Remove unescaping of LaTeX code
Partially escaping strings for LaTeX code created more problems
than it solved. So now message (and other property) strings
are taken as-is. That means the following combinations are
taken as LilyPond escape sequences:
- \n (newline)
- \t (tab)
- \" (quotation mark).
If one of these is to appear in the output it has to be escaped
like \\n \\t \\".
The last one is unlikely, but the first two may be needed for
LaTeX commands, e.g. \\noindent.