comptools: Factor out shared code in __init__
This is a good example for library initialization.
There is shared code between partial-compilation and
conditional-breaks, which is loaded upon library initialization.

the *library* maintains the list of break sets and provides
the interface for that. This also means that when both modules
are loaded they share that list and should take care to use
either the same or explicitly different break sets.