test: improve version string regex
It was necessary to change at least the last part of the match
because the third part of the version string can well be one
or two digits long.

The commit modifies all three string segments to be at least
one digit long because in theory also the major and minor
versions might be of any length.