test: Change test collection strategy
Instead of specifying explicitly the files that should be used as tests,
now the `simple_tests.py` script will look for all files in
`usage-examples` directories and use them as tests.

A developer can make the script ignore some specific files by listing
them in a `.simple-test-excludes` file in the appropriate
`usage-examples` directory.