gridly: Simplify default cell creation
When using a cell that has no music inside, but for which a template has
been defined, use the template's contents instead of building an ad hoc
piece of 'SequentialMusic. This is possible since the <template> part no
longer contains marks and tempo changes (for which we have to use a
dedicated `marks` part).

All the code related to the `make-skips` function has been removed,
since it is no longer needed.