test: define results dir (to be discussed)
Previously test results were written to the working directory,
which is not good.
This commit defines a directory test/results where everything goes.

However, I'm not sure about it because that would mean we would
impose the requirement of unique file names throughout the whole
openLilyLib directory structure.

So the question is: Should results go beside the test files?
This would easily be doable by dynamically generating the -o
parameter.