test: store test results in shadow hierarchy
Now a test produces two shadow hierarchies below `test/results`,
one for each tested version (stable and devel).

Before tests are started the shadow directory is cleared so
they are not intended for manual editing.
After the test ran successfully the directory/ies is/are
available for picking result files for documentation generation.