test: Installation script for Travis
This commit adds an installation script that is run before the
tests are performed on Travis. (Temporarily the execution of
the tests themselves are disabled by this commit - but only for
concentrating on the Travis integration of the installation).

- LilyPond versions are now configured using a dedicated file.
- Two LilyPond installations are cached between builds
- When the script finds an installed LilyPond that isn't
  requested (presumably the first build after a version upgrade
  has become available) it is removed from cache.
- When a requsted Lilypond version isn't present in the cache
  it is downloaded and installed.

If run in CI mode (i.e. on Travis) LilyPond is downloaded and
installed, but only when there is no relevant cached version.

If LilyPond versions to be used have changed, existing older
installations are removed (hopefully ...)