gridly: cache segment-range in gridCompileCell
This commit removes a side-effect of \gridCompileCell.
Previously this command modified the overall segment range
so if a segment was compiled along with any score making use of the
grid that score was limited to the single segment of the cell.

Now the segment-range option is cached and re-applied at the end of
the command.