gridly: Formatting for gridCompileCell filenames
This commit is a follow-up of 814e1958f025ca68be72bf4bbe233f60647a1050.
Now the "segment id" part of files compiled with gridCompileCell is
padded with zeroes to the length of the longest segment-id string. Thus
the cell of "soprano 4" in a grid with 100 segments, when compiled with
gridCompileCell, will produce the following pdf file

  soprano-004.pdf