test: don't test .ily files
In test directories only .ly files are now collected.
It should be an extremely rare case that we want to pass an .ily
file to LilyPond. And in these cases the file can still be added
manually.

The is_lilypond_file() check in is_runnable_file()
can be dropped because:
- with automatically collected files it has already
  been done
- with explicitly named tests we don't want to impose
  anything, it's up to the library maintainer.