oll: Fix location-extract-path
Strips the trailing slash from the location path.
I'm not completely sure if that will have side effects on existing code,
but the automated tests do work.

Until now the root-path did have a trailing slash, which caused any of
the split-path based processes to erroneously add a double slash in
resulting paths.