stylesheets: Let \useNotationFont read \with
if a \with block is passed as first argument it is parsed for given
options. The first option to be recognized is brace:
\with {
  brace = font-name % or "font-name"
}
sets the brace font to that name.
If it is omitted the brace font is the same as the notation font.
brace = none will default the brace to Emmentaler (which may make sense
when there is no brace font available).