stylesheets: load font and (default) styles
\useNotationFont now configures the font and loads the font and a
stylesheet it in one command.
It uses an included file for loading the font (as long as we don't have
a solution for that issue), and includes a separate style sheet.
Any available font should have at least a -default stylesheet associated
(which also can be empty but must exist as a file).