font-install: Prepare for handling multiple lilys
The LilyPond font root directory is now stored in a list, and
when checking/updating the links this list is iterated.

Currently this list is populated with the single determined
LilyPond installation, but it is checked that it also works
with multiple installations.
So the next step is to create a user interface to pass multiple
LilyPonds to the script (maybe the best way is to use a file
for that?)