test: Fix caching issues (#111)
This commit deals with the problems of caching LilyPond versions between
Travis-CI builds reported in #111. Specifically, it turned out that
there is not a separate cache per build-matrix entry (since LilyPond is
not a supported language), but only a dedicated cache per branch.

The issue was then that only one version of LilyPond was actually used
to run the tests, since the installation path was identical for all
versions.

The solution introduced in this commit is to have an explicit cache
management mechanism:

 - Each LilyPond version is installed in its own folder under the
   $HOME/.lilypond cache directory, for instance

     $HOME/.lilypond/linux-64/2.18.2-1

 - Every time a cached LilyPond version is used, the timestamp is
   written to a file in the cache, for later retrieval
 - After all the tests have been run, the cache is cleaned: LilyPond
   versions that have not been used for more than one day (this is
   configurable) are removed from the cache. Note that since this
   cleanup phase is run after the tests have been run, all the LilyPond
   versions that have actually been used in the last test run are kept
   in the cache