Fix non-existent default style bug
When an explicitly requested stylesheet or a default stylesheet
for a font isn't found we now only issue a warning instead of
letting LilyPond run into a file-not-found error.

This has become necessary as people may be in the posession of
fonts that are not part of the official collection (and that
therefore don't have a default stylesheet in OLL).