module.scm: code reformatting
Jan-Peter, this is your project. But in order for others to read and understand the code, I reformatted it according (mostly) to the guidelines in <http://community.schemewiki.org/?scheme-style>, which I find quite convincing. A main issue which can’t always be solved here (without moving code around) is to avoid long lines with more than ~80 characters.