core: adapt to new 'parser' syntax
In LilyPond 2.19.22 a significant change has been applied, making it
possible to write much simpler code because the 'parser location'
arguments are not required anymore.

However, maintaining Lily code for older and newer versions is not really
trivial that way.

This commit (hopefully) fixes #125, at least for any cases that are already inside the 'ly' folder structure.