Hack to fix (???) the tablature-microtones issue with 2.19.31
As described in the previous commit sometimes the element taken as
"string number" doesn't contain a string but a markup function in 2.19.31

This commit *prevents* the failure by checking for this condition and
forcing the string number to be "1".

I'm sure this is no proper cure, but I wanted to have this online for
further investigation.