compoundSlur: Change option checking behaviour
Options are now checked against a list of default values and predicates.
Missing options are still set to default values, but values with
mismatching predicates issue a warning and fall back to default value.

(Mixed into this is the display of the original slur and the
control points, which can now be selected independently)