Make end-angle behaviour consistent
I think it was no good idea to have the end angle mirror its value.
Now the end angle is treated as if continuing to the right.
While this may seem strange from the slur's perspective it is more
consistent to think about the angle when going through the inflections.